Rework scripts directory
