+++ gpt/generate.js
@@ -1,5 +1,5 @@
-const fs = require('fs')
-const { glob } = require('glob')
+import fs from 'node:fs'
+import { glob } from 'glob'
 
 const outputPath = './gpt/db.json'
 const packages = []
