+++ package.json
@@ -5,6 +5,8 @@
     "type": "git",
     "url": "https://github.com/TanStack/router.git"
   },
+  "packageManager": "pnpm@8.12.1",
+  "type": "module",
   "scripts": {
     "clean": "pnpm --filter \"./packages/**\" run clean",
     "preinstall": "node -e \"if(process.env.CI == 'true') {console.log('Skipping preinstall...'); process.exit(1)}\" || npx -y only-allow pnpm",
@@ -43,6 +45,7 @@
     "@rollup/plugin-commonjs": "^25.0.7",
     "@rollup/plugin-node-resolve": "^15.2.3",
     "@rollup/plugin-replace": "^5.0.5",
+    "@tanstack/config": "^0.2.1",
     "@tanstack/react-router": "^1.0.0",
     "@testing-library/react": "^14.1.2",
     "@testing-library/react-hooks": "^8.0.1",
